paystack_project/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
